+++ Example/Backpack/FeatureStories/Groups/CardGroups.swift
@@ -40,9 +40,18 @@ struct CardGroupsProvider {
         )
     }
     
+    private var swiftUI: CellDataSource {
+        PresentableCellDataSource.custom(
+            title: "SwiftUI",
+            customController: { ContentUIHostingController(CardExampleView()) },
+            showPresentable: showPresentable
+        )
+    }
+    
     func groups() -> [Components.Group] {
         SingleGroupProvider(
             cellDataSources: [
+                swiftUI,
                 presentableCard("Default") { _ in },
                 presentableCard("Button") { $0.configuration = .button },
                 presentableCard("Link") { $0.configuration = .link },